% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeIpRotation.R
\name{completeIpRotation}
\alias{completeIpRotation}
\title{completeIpRotation}
\usage{
completeIpRotation(projectId, zone, clusterId)
}
\arguments{
\item{projectId}{The Google Developers Console project ID or project number.}

\item{zone}{The name of the Google Compute Engine zone in which the cluster resides.}

\item{clusterId}{The name of the cluster.}
}
\description{
Completes master IP rotation.
}
\examples{
completeIpRotation(projectId = "myProjectId", zone = "europe-west1-d", clusterId = "cluster-1")
}
\keyword{projects.locations.getServerConfig}
