% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCluster.R
\name{createCluster}
\alias{createCluster}
\title{createCluster}
\usage{
createCluster(projectId, zone, clusterSpec = NULL)
}
\arguments{
\item{projectId}{The Google Developers Console project ID or project number.}

\item{zone}{The name of the Google Compute Engine zone in which the cluster resides.}

\item{clusterSpec}{A cluster specification. If no specification is provided, a 3 node, 100Gb disk, n1-standard-1 cluster will be created with the name "r-cluster". See here for specification: https://cloud.google.com/kubernetes-engine/docs/reference/rest/v1beta1/projects.zones.clusters}
}
\description{
Creates a cluster, consisting of the specified number and type of Google Compute Engine instances.
By default, the cluster is created in the project's default network.
One firewall is added for the cluster. After cluster creation, the cluster creates routes for each node to allow the containers on that node to communicate with all other instances in the cluster.
Finally, an entry is added to the project's global metadata indicating which CIDR range is being used by the cluster.
}
\examples{
createCluster(projectId = "myProjectId", zone = "europe-west1-d")
}
\keyword{projects.zones.clusters.create}
