% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteCluster.R
\name{deleteCluster}
\alias{deleteCluster}
\title{deleteCluster}
\usage{
deleteCluster(projectId, zone, clusterId)
}
\arguments{
\item{projectId}{The Google Developers Console project ID or project number.}

\item{zone}{The name of the Google Compute Engine zone in which the cluster resides.}

\item{clusterId}{The name of the cluster to delete.}
}
\description{
Deletes the cluster, including the Kubernetes endpoint and all worker nodes.
Firewalls and routes that were configured during cluster creation are also deleted.
Other Google Compute Engine resources that might be in use by the cluster (e.g. load balancer resources) will not be deleted if they weren't present at the initial create time.
}
\examples{
deleteCluster(projectId = "myProjectId", zone = "europe-west1-d", clusterId = "r-cluster")
}
\keyword{projects.zones.clusters.delete}
