% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCluster.R
\name{getCluster}
\alias{getCluster}
\title{getCluster}
\usage{
getCluster(projectId, zone, clusterId)
}
\arguments{
\item{projectId}{The Google Developers Console project ID or project number.}

\item{zone}{The name of the Google Compute Engine zone in which the cluster resides.}

\item{clusterId}{The name of the cluster to retrieve.}
}
\description{
Gets the details of a specific cluster.
}
\examples{
getCluster(projectId = "myProjectId", zone = "europe-west1-d", clusterId = "cluster-1")
}
\keyword{projects.locations.clusters.get}
