% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getServerConfig.R
\name{getServerConfig}
\alias{getServerConfig}
\title{getServerConfig}
\usage{
getServerConfig(projectId, zone)
}
\arguments{
\item{projectId}{The Google Developers Console project ID or project number.}

\item{zone}{The name of the Google Compute Engine zone in which the cluster resides.}
}
\description{
Returns configuration info about the Kubernetes Engine service.
}
\examples{
getServerConfig(projectId = "myProjectId", zone = "europe-west1-d")
}
\keyword{projects.locations.getServerConfig}
