% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAddons.R
\name{setAddons}
\alias{setAddons}
\title{setAddons}
\usage{
setAddons(projectId, zone, clusterId, addonsConfig)
}
\arguments{
\item{projectId}{The Google Developers Console project ID or project number.}

\item{zone}{The name of the Google Compute Engine zone in which the cluster resides.}

\item{clusterId}{The name of the cluster to retrieve.}
}
\description{
Sets the addons of a specific cluster.
}
\examples{
setAddons(projectId = "myProjectId",
  location = "europe-west1-d",
  clusterId = "cluster-1",
  addonsConfig = list(
    httpLoadBalancing = list(
      disabled = FALSE
    )
  )
)
}
\keyword{projects.zones.clusters.addons}
