% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClusterLocation.R
\name{setClusterLocation}
\alias{setClusterLocation}
\title{setClusterLocation}
\usage{
setClusterLocation(projectId, zone, clusterId, locations)
}
\arguments{
\item{projectId}{The Google Developers Console project ID or project number.}

\item{zone}{The name of the Google Compute Engine zone in which the cluster resides.}

\item{clusterId}{The name of the cluster to upgrade.}

\item{locations}{The desired list of Google Compute Engine locations in which the cluster's nodes should be located. Changing the locations a cluster is in will result in nodes being either created or removed from the cluster, depending on whether locations are being added or removed. This list must always include the cluster's primary zone.}
}
\description{
Sets the locations of a specific cluster.
}
\examples{
setClusterLocation(projectId = "myProjectId",
                   zone = "europe-west1-d",
                   clusterId = "r-cluster",
                   locations = c("europe-west1-d",
                                 "europe-west1-b"))
}
\keyword{projects.zones.clusters.locations}
