% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k8s_auth.R
\name{k8s_auth}
\alias{k8s_auth}
\title{k8s_auth}
\usage{
k8s_auth(json = NULL, service_account_json = NULL)
}
\arguments{
\item{json}{OPTIONAL Path to your JSON client_secret credentials.}

\item{service_account_json}{OPTIONAL Path to your JSON service account credentials}
}
\description{
Calls googleAuthR to authenticate with the Google Cloud Platform
}
\examples{
k8s_auth()
k8s_auth(json = "client_secret.json")
k8s_auth(service_account_json = "service_account.json")
}
\keyword{k8s_auth}
